feat: keep build artifacts in different directories for each profile

+++ ci/run.sh
@@ -2,15 +2,46 @@
 
 set -ex
 
-cargo install --path .
+main() {
+  setup
 
-# Test with supported target
-cargo sysroot --target arm-unknown-linux-gnueabihf sysroot --verbose
-tree sysroot/lib
-rm -r sysroot
+  test_supported_target
+  test_custom_target
+}
 
-# Test with custom target
-cat >custom.json <<EOF
+die() {
+  echo "$@" 1>&2
+  exit 1
+}
+
+setup() {
+  cargo install --path .
+}
+
+test_supported_target() {
+  local triple=arm-unknown-linux-gnueabihf
+
+  # build sysroot in debug mode
+  cargo sysroot --target $triple sysroot --verbose
+
+  # check that libcore was built in debug mode
+  [ -e sysroot/debug/lib/rustlib/$triple/lib/libcore.rlib ] || die
+
+  # build sysroot in release mode
+  cargo sysroot --target $triple sysroot --release --verbose
+
+  # check that libcore was built in release mode
+  [ -e sysroot/release/lib/rustlib/$triple/lib/libcore.rlib ] || die
+
+  # check that the debug mode libcore still exists
+  [ -e sysroot/debug/lib/rustlib/$triple/lib/libcore.rlib ] || die
+
+  # clean up
+  rm -r sysroot
+}
+
+test_custom_target() {
+  cat >custom.json <<EOF
 {
   "arch": "arm",
   "llvm-target": "thumbv7m-none-eabi",
@@ -19,12 +50,32 @@ cat >custom.json <<EOF
   "target-pointer-width": "32"
 }
 EOF
-cp custom.json ..
-cargo sysroot --target ../custom.json sysroot --verbose
-# check that the original spec file is still there
-[ -e ../custom.json ] || exit 1
-# check that the copied spec file was removed
-[ -e sysroot/src/libcore/custom.json ] && exit 1
-tree sysroot/lib
-rm -r sysroot
-rm custom.json
+
+  # test --target triple
+  cargo sysroot --target custom sysroot --verbose
+
+  # confirm existence of build artifacts
+  [ -e sysroot/debug/lib/rustlib/custom/lib/libcore.rlib ] || die
+
+  # clean up
+  rm -r sysroot
+
+  # test --target path/to/triple.json
+  cp custom.json ..
+  cargo sysroot --target ../custom.json sysroot --verbose
+
+  # confirm existence of build artifacts
+  [ -e sysroot/debug/lib/rustlib/custom/lib/libcore.rlib ] || die
+
+  # check that the original spec file is still there
+  [ -e ../custom.json ] || die
+
+  # check that the copied spec file was removed
+  [ -e sysroot/src/libcore/custom.json ] && die
+
+  # clean up
+  rm ../custom.json
+  rm -r sysroot
+}
+
+main
